직원 리스트, 상세보기 추가, 직원게시판 로그인직원은 본인 댓글만 삭제할 있도록 수정
StaffController 추가, staffService, staffMapper 수정

+++ sakila/src/main/java/com/gd/sakila/controller/CommentController.java
@@ -1,5 +1,7 @@
 package com.gd.sakila.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.gd.sakila.service.CommentService;
 import com.gd.sakila.vo.Comment;
+import com.gd.sakila.vo.Staff;
 
 import lombok.extern.slf4j.Slf4j;
 @Slf4j//log 각체 생성
@@ -20,17 +23,28 @@ public class CommentController {
 	
 	@GetMapping("/removeComment")
 	public String removeComment(@RequestParam(value = "commentId", required = true)int commentId,
-								@RequestParam(value = "boardId", required = true)int boardId) {
+								@RequestParam(value = "boardId", required = true)int boardId,
+								@RequestParam(value = "username",required = true)String username,
+								HttpSession session) {
+		
 		//required = true --> 필수 입력 사항
 		log.debug("▶▶▶▶▶▶ removeComment comentId param: "+commentId);
+		log.debug("▶▶▶▶▶▶ removeComment username param: "+username);
 		log.debug("▶▶▶▶▶▶ removeComment boardId param: "+boardId);
 		
-		//댓글 삭제 서비스 호출
-		int row = commentService.removeCommet(commentId);
-		log.debug("▶▶▶▶▶▶ removeComment row: "+row);
+		//댓글을 쓴 본인만 삭제 할 수 있도록 바꿈
 		
+		Staff staff = (Staff)session.getAttribute("loginStaff");
+		log.debug("▶▶▶▶▶▶ removeComment staff: "+staff);
+		
+		//댓글 삭제 서비스 호출
+		if(staff.getUsername().equals(username)) {
+			int row = commentService.removeCommet(commentId);
+			log.debug("▶▶▶▶▶▶ removeComment row: "+row);
+		}		
 		return "redirect:/admin/getBoardOne?boardId="+boardId;
 	}
+	
 	
 	@PostMapping("/addComment")
 	public String addComment(Comment comment) {
